refactor(webdto): MoveRequest/MoveResponse 패키지 이동
